Move SearchParams to History providers
