+++ client/src/components/History/ContentFilters.vue
@@ -31,7 +31,7 @@
 </template>
 
 <script>
-import { SearchParams } from "./model/SearchParams";
+import { SearchParams } from "components/providers/History/SearchParams";
 import DebouncedInput from "components/DebouncedInput";
 
 export default {
