+++ client/src/components/History/providers/HistoryContentProvider/HistoryContentProvider.test.js
@@ -2,7 +2,7 @@ import { shallowMount } from "@vue/test-utils";
 import { watchUntil, getLocalVue } from "jest/helpers";
 import { sameVueObj, payloadChange } from "../../test/providerTestHelpers";
 import HistoryContentProvider from "./HistoryContentProvider";
-import { SearchParams } from "../../model";
+import { SearchParams } from "components/providers/History/SearchParams";
 import { bulkCacheContent, wipeDatabase } from "components/providers/History/caching";
 import { defaultPayload } from "../ContentProvider";
 import { serverContent, testHistory, testHistoryContent } from "../../test/testHistory";
