+++ client/src/components/History/test/testHistory.js
@@ -1,4 +1,5 @@
-import { History, SearchParams } from "../model";
+import { History } from "../model";
+import { SearchParams } from "components/providers/History/SearchParams";
 import rawHistory from "./json/History.json";
 import rawHistoryContent from "./json/historyContent.json";
